android15-6.6 July 2024 release 1

Artifacts:
  https://ci.android.com/builds/submitted/12057461/kernel_aarch64/latest

* tag 'android15-6.6-2024-07_r1': (187 commits)
  FROMLIST: mm: fix incorrect vbq reference in purge_fragmented_block
  ANDROID: sched/psi: disable the privilege check if CONFIG_DEFAULT_SECURITY_SELINUX is enabled
  UPSTREAM: mm/vmalloc: fix vmalloc which may return null if called with __GFP_NOFAIL
  ANDROID: ABI: Update symbol list for OPLUS
  ANDROID: ABI: Update oplus symbol list
  ANDROID: Revert^3 "ANDROID: Enable Rust Binder Module"
  ANDROID: abi fixups for HCI_AMP removal changes
  UPSTREAM: Bluetooth: HCI: Remove HCI_AMP support
  ANDROID: Update the ABI symbol list
  ANDROID: softirq: Add EXPORT_SYMBOL_GPL for softirq and tasklet
  ANDROID: fix the --entry linker flag for kselftest
  ANDROID: abi_gki_aarch64_qcom: update abi symbols
  Revert^2 "ANDROID: Enable Rust Binder Module"
  ANDROID: GKI: Update qcom symbol list
  ANDROID: ABI: update symbol list for honor
  ANDROID: vendor_hooks: add vendor hook for bpf jit
  ANDROID: GKI: update symbol list file for xiaomi
  ANDROID: abi_gki_aarch64_vivo: Update symbol list
  ANDROID: vendor_hooks: add hook in alloc_and_link_pwqs()
  ANDROID: GKI: export symbols
  ...

+++ drivers/dma-buf/heaps/carveout-heap.c
@@ -16,7 +16,6 @@
 #include <linux/highmem.h>
 #include <linux/dma-buf.h>
 #include <linux/dma-heap.h>
-#include <linux/dma-heap-carveout.h>
 
 struct carveout_dma_heap {
 	struct dma_heap *heap;
@@ -258,8 +257,8 @@ static const struct dma_buf_ops carveout_dma_heap_buf_ops = {
 
 static struct dma_buf *carveout_dma_heap_allocate(struct dma_heap *heap,
 						  unsigned long len,
-						  unsigned long fd_flags,
-						  unsigned long heap_flags)
+						  u32 fd_flags,
+						  u64 heap_flags)
 {
 	struct carveout_dma_heap *carveout_dma_heap = dma_heap_get_drvdata(heap);
 	struct carveout_dma_heap_buffer *buffer;
